+++ Day06-09/PhungDacNhatMinh/day4/day4.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <script src="./day4.js"></script>
+    <script type="Module" src="./day4.js"></script>
 </body>
 </html>
