+++ Day06-09/PhungDacNhatMinh/day4/modules/convertToArray.js
@@ -1,0 +1,5 @@
+import convertToString from "./convertToString.js";
+ function convertToArray() {
+    return (convertToString().split(","));
+  }
+  export default convertToArray;
