+++ Day06-09/PhungDacNhatMinh/day4/modules/convertToString.js
@@ -1,0 +1,6 @@
+import firstName from "./firstName.js";
+ function convertToString() {
+    return (firstName().join());
+  }
+
+  export default convertToString;
