+++ Day06-09/PhungDacNhatMinh/day4/modules/data.js
@@ -1,0 +1,180 @@
+ const arrStudentEdsoLabs = [
+    {
+      id: 1,
+      name: "Nguyễn Văn Sơn",
+      group: {
+        position: "Thành viên",
+        groupId: "1",
+      },
+    },
+    {
+      id: 2,
+      name: "Nguyễn Hữu Ánh",
+      group: {
+        position: "Thành viên",
+        groupId: "1",
+      },
+    },
+    {
+      id: 3,
+      name: "Trần Mạnh Quân",
+      group: {
+        position: "Trưởng nhóm",
+        groupId: "4",
+      },
+    },
+    {
+      id: 4,
+      name: "Hà Quốc Tuấn",
+      group: {
+        position: "Trưởng nhóm",
+        groupId: "3",
+      },
+    },
+    {
+      id: 5,
+      name: "Hoàng Ngọc Thành",
+      group: {
+        position: "Thành viên",
+        groupId: "1",
+      },
+    },
+    {
+      id: 6,
+      name: "Vũ Thị Thu Hà",
+      group: {
+        position: "Thành viên",
+        groupId: "2",
+      },
+    },
+    {
+      id: 7,
+      name: "Phan Văn Trung",
+      group: {
+        position: "Thành viên",
+        groupId: "2",
+      },
+    },
+    {
+      id: 8,
+      name: "Nguyễn Cao Hoàng",
+      group: {
+        position: "Thành viên",
+        groupId: "2",
+      },
+    },
+    {
+      id: 9,
+      name: "Phùng Đắc Nhật Minh",
+      group: {
+        position: "Trưởng nhóm",
+        groupId: "5",
+      },
+    },
+    {
+      id: 10,
+      name: "Lê Việt Dũng",
+      group: {
+        position: "Trưởng nhóm",
+        groupId: "1",
+      },
+    },
+    {
+      id: 11,
+      name: "Đỗ Chí Dũng",
+      group: {
+        position: "Thành viên",
+        groupId: "2",
+      },
+    },
+    {
+      id: 12,
+      name: "Trần Công Tâm",
+      group: {
+        position: "Thành viên",
+        groupId: "3",
+      },
+    },
+    {
+      id: 13,
+      name: "Trương Thanh Tùng",
+      group: {
+        position: "Thành viên",
+        groupId: "3",
+      },
+    },
+    {
+      id: 14,
+      name: "Tạ Đức Chiến",
+      group: {
+        position: "Thành viên",
+        groupId: "3",
+      },
+    },
+    {
+      id: 15,
+      name: "Nguyễn Trọng Vinh",
+      group: {
+        position: "Thành viên",
+        groupId: "3",
+      },
+    },
+    {
+      id: 16,
+      name: "Ngô Chung Á Âu",
+      group: {
+        position: "Thành viên",
+        groupId: "4",
+      },
+    },
+    {
+      id: 17,
+      name: "Trần Thị Khánh Linh",
+      group: {
+        position: "Trưởng nhóm",
+        groupId: "2",
+      },
+    },
+    {
+      id: 18,
+      name: "Phan Tiến Thành",
+      group: {
+        position: "Thành viên",
+        groupId: "4",
+      },
+    },
+    {
+      id: 19,
+      name: "Đỗ Văn Huy",
+      group: {
+        position: "Thành viên",
+        groupId: "4",
+      },
+    },
+    {
+      id: 20,
+      name: "Nguyễn Trung Đức",
+      group: {
+        position: "Thành viên",
+        groupId: "5",
+      },
+    },
+    {
+      id: 21,
+      name: "Nguyễn Trung Nam",
+      group: {
+        position: "Thành viên",
+        groupId: "5",
+      },
+    },
+    {
+      id: 22,
+      name: "Trần Quốc Toàn",
+      group: {
+        position: "Thành viên",
+        groupId: "5",
+      },
+    },
+  ];
+
+  export default arrStudentEdsoLabs;
